--- messaging/messages.go
@@ -1,6 +1,0 @@
-package messaging
-
-type Sendable struct {
-	Ids  []string
-	Data []byte
-}
